requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
openai>=1.3.7
tenacity>=8.2.3
google-api-python-client>=2.108.0
google-auth>=2.23.4
hubspot-api-client>=9.0.0
python-dotenv>=1.0.0
deepdiff>=6.7.1
httpx>=0.25.2
tiktoken>=0.5.2 